move event manager setup back to core, remove ref to global EVENT_MANAGER and clean up event manager functions
